# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An `API` for cross-platform custom orchestration of execution steps\nwithout any third-party dependencies. Based on `DAG`, it implements the scheduling\nfunction of sequential execution of dependent steps and concurrent execution of\nnon-dependent steps. \u003cbr /\u003e\u003cbr /\u003e It provides `API` remote operation mode, batch\nexecution of `Shell` , `Powershell` , `Python` and other commands, and easily\ncompletes common management tasks such as running automated operation and maintenance\nscripts, polling processes, installing or uninstalling software, updating applications,\nand installing patches.",
        "title": "DAG Flow",
        "contact": {
            "name": "dagflow",
            "url": "https://github.com/busyster996/dagflow/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/busyster996/dagflow/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/task/{task}/workspace": {
            "get": {
                "description": "获取目录列表或下载指定文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作目录"
                ],
                "summary": "列表或下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SFileListRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "上传文件或目录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作目录"
                ],
                "summary": "上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定目录或文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作目录"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Code": {
            "type": "integer",
            "enum": [
                0,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeRunning",
                "CodeFailed",
                "CodeNoData",
                "CodePending",
                "CodePaused",
                "CodeSkipped"
            ]
        },
        "types.SBase-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_SFileListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.SFileListRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "types.SFileListRes": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SFileRes"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.SFileRes": {
            "type": "object",
            "properties": {
                "isDir": {
                    "type": "boolean"
                },
                "modTime": {
                    "type": "integer"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
    }
}